requirements.txt
imbalanced-learn==0.11.0
imblearn==0.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kaggle==1.6.3
keras==2.15.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.23.5
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==1.5.3
Pillow==10.1.0
plot-keras-history==1.1.38
plotly==5.18.0
plotly-resampler==0.9.1
scikit-base==0.5.2
scikit-image==0.22.0
scikit-learn==1.2.2
scikit-plot==0.3.7
seaborn==0.13.0
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tqdm==4.66.1

